fix journal stop bug
